@@ -1,12 +1,12 @@
-FactoryGirl.define do
+FactoryBot.define do
   factory :user do
-    confirmed_at Time.now
-    name "Test User"
-    email "test@example.com"
-    password "please123"
+    confirmed_at { Time.now }
+    name { "Test User" }
+    email { "test@example.com" }
+    password { "please123" }
 
     trait :admin do
-      role 'admin'
+      role { 'admin' }
     end
 
   end
