@@ -16,8 +16,8 @@ feature 'User edit', :devise do
   #   When I change my email address
   #   Then I see an account updated message
   scenario 'user changes email address' do
-    user = FactoryGirl.create(:user)
-    login_as(user, :scope => :user)
+    user = FactoryBot.create(:user)
+    signin(user.email, user.password)
     visit edit_user_registration_path(user)
     fill_in 'Email', :with => 'newemail@example.com'
     fill_in 'Current password', :with => user.password
@@ -31,9 +31,9 @@ feature 'User edit', :devise do
   #   When I try to edit another user's profile
   #   Then I see my own 'edit profile' page
   scenario "user cannot cannot edit another user's profile", :me do
-    me = FactoryGirl.create(:user)
-    other = FactoryGirl.create(:user, email: 'other@example.com')
-    login_as(me, :scope => :user)
+    me = FactoryBot.create(:user)
+    other = FactoryBot.create(:user, email: 'other@example.com')
+    signin(me.email, me.password)
     visit edit_user_registration_path(other)
     expect(page).to have_content 'Edit User'
     expect(page).to have_field('Email', with: me.email)
